requirements.txt
openai
pandas
tiktoken
scikit-learn
python-dotenv
matplotlib
plotly
langchain
pypdf
faiss-cpu